+++ processing/provider.py
@@ -3,6 +3,10 @@ from qgis.core import QgsProcessingProvider
 from qgis.PyQt.QtGui import QIcon
 from ThreeDiToolbox.processing.dwf_calculation_algorithm import DWFCalculatorAlgorithm
 from ThreeDiToolbox.processing.threedidepth_algorithm import ThreediDepth
+from ThreeDiToolbox.processing.schematisation_algorithms import (
+    CheckSchematisationAlgorithm,
+    MigrateAlgorithm
+)
 
 
 class ThreediProvider(QgsProcessingProvider):
@@ -11,8 +15,8 @@ class ThreediProvider(QgsProcessingProvider):
     def loadAlgorithms(self, *args, **kwargs):
         self.addAlgorithm(ThreediDepth())
         self.addAlgorithm(DWFCalculatorAlgorithm())
-        # add additional algorithms here
-        # self.addAlgorithm(MyOtherAlgorithm())
+        self.addAlgorithm(CheckSchematisationAlgorithm())
+        self.addAlgorithm(MigrateAlgorithm())
 
     def id(self, *args, **kwargs):
         """The ID of your plugin, used for identifying the provider.